requirements.txt
# testing, formatting, linting
mypy
black
isort
flake8

pydantic~=2.10.5
python-dotenv~=1.0.1
PyYAML~=6.0.2
langchain~=0.3.15
langchain-community
langchain-chroma
langchain-openai
langchain-experimental
pytest~=8.3.4
pytest-asyncio
streamlit~=1.41.1